[docs](kt-kernel): consolidate and simplify the Ascend deployment scripts
Review asked why a docs PR carries this many scripts and lines, and asked for a
per-script justification. Reworked to the intended scope: the scripts package
commands, the tutorial tells you which commands to run.

  16 files, 3074 lines -> 8 files, 2003 lines

tools/ascend_dsv4/  11 files 1370 lines -> 5 files 1057

  setup.sh   absorbs install_python_deps.sh, build_kt_kernel.sh,
             build_sgl_kernel_npu.sh, build_cann_ops.sh, convert_mxfp4_gguf.sh
             and preflight.sh as the subcommands deps / kt-kernel / sgl-kernel /
             cann-ops / gguf / check, plus `all`. These were always run in this
             order and shared one configuration surface.
  verify.sh  absorbs dsv4_chat.py as `verify.sh chat`.

tools/mxfp4_gguf/  4 files 718 lines -> 2 files 606

  convert_mxfp4_gguf.py  `layer` (one layer) + `batch` (fan out across processes)
  verify_mxfp4_gguf.py   `layer` (bit-exactness) + `set` (whole directory)

doc/  986 lines -> 340

  Restructured as a linear walkthrough: start the container, get the code,
  configure ~/dsv4.env, download the weights, `setup.sh all`, launch, verify,
  use it. Design rationale and the long troubleshooting catalogue are gone; a
  configuration reference table, the long-prompt settings, the measured results
  and a five-row troubleshooting table remain.

Explanatory comments are removed from the scripts throughout. Executable logic is
carried over verbatim -- the merges only change the argparse entry points, the
subprocess targets (now the same file) and the sibling imports.

Verification: all six scripts pass `bash -n` / `py_compile`; the four Python
subcommands were exercised for their CLI surface; and the converter's own unit
check (pure-numpy repack -> dequant equality) was run on an Ascend host against
the original four files and against the merged pair, reporting the same result
either way -- "repack lossless, 896 elements bit-exact". The build steps
themselves are unchanged code and were not re-run.
